chore: Fix urlsafe_base64/1 generation
The test assumes 22 byte default, but had 32 because it was double-base64 encoding (once
regular, and once URL-safe). Without the double encoding, the tests fail because they were
expecting 32 (then 24), but the intent was URL-safe base encoding without padding.

Originally based on patricksrobertson/secure_random.ex#13, but fixes the tests and
padding.